[Feat]메뉴-> gemini 연동버전으로

+++ Projects/Soonsu/Model/DailySales.swift
@@ -10,7 +10,7 @@ import Foundation
 /// 하루 치 매출/원가 합산 정보
 struct DailySales: Codable, Hashable {
     let revenue: Int
-//    let materialCost: Int
-    let materialCost: Double
+    let materialCost: Int
+//    let materialCost: Double
     let items: [SoldItem]
 }
